Update NuGet workflow and project files
- Updated the `dotnet-version` in the NuGet workflow to use version 8.x
- Updated the project properties in `Directory.build.props` with new metadata values
- Removed unused project files: `Tycho.JsonSerializer.NewtonsoftJson.csproj` and `Tycho.JsonSerializer.SystemTextJson.csproj`

These changes improve the NuGet packaging process and update project metadata.
