--- Tycho.Benchmarks/Program.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using BenchmarkDotNet.Running;
-
-namespace Tycho.Benchmarks;
-
-public static class Program
-{
-    public static void Main(string[] args)
-    {
-        var summary = BenchmarkRunner.Run<Benchmarks.Insertion>();
-    }
-}
