--- Tycho.Benchmarks/Tycho.Benchmarks.csproj
@@ -1,40 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0</TargetFramework>
-    <Configurations>Debug;Release;Encrypted</Configurations>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' " />
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType></DebugType>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Encrypted|AnyCPU' ">
-    <IntermediateOutputPath>obj\Release</IntermediateOutputPath>
-    <DebugType></DebugType>
-    <Optimize>true</Optimize>
-    <DefineConstants>ENCRYPTED;</DefineConstants>
-    <WarningLevel>4</WarningLevel>
-  </PropertyGroup>
-  <ItemGroup>
-    <None Remove="Benchmarks\" />
-    <None Remove="BenchmarkDotNet" />
-    <None Remove="sqlite-net-pcl" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Benchmarks\">
-    </Folder>
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="BenchmarkDotNet" Version="0.13.7" />
-    <PackageReference Include="sqlite-net-pcl" Version="1.8.116" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Tycho\Tycho.csproj" />
-    <ProjectReference Include="..\Tycho.JsonSerializer\Tycho.JsonSerializer.csproj" />
-    <ProjectReference Include="..\Tycho.JsonSerializer.NewtonsoftJson\Tycho.JsonSerializer.NewtonsoftJson.csproj" />
-    <ProjectReference Include="..\Tycho.JsonSerializer.SystemTextJson\Tycho.JsonSerializer.SystemTextJson.csproj" />
-  </ItemGroup>
-</Project>
