--- Tycho.JsonSerializer.SystemTextJson/Tycho.JsonSerializer.SystemTextJson.csproj
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project Sdk="MSBuild.Sdk.Extras/3.0.22">
-	<PropertyGroup>
-		<TargetFrameworks>netstandard2.1;NET7.0</TargetFrameworks>
-		<LangVersion>Latest</LangVersion>
-		<PackageId>Tycho.JsonSerializer.SystemTextJson</PackageId>
-		<Authors>Eight-Bot, Inc.</Authors>
-		<Copyright>2023 Eight-Bot</Copyright>
-		<PackageIcon>icon.png</PackageIcon>
-		<Owners>Eight-Bot, Inc.</Owners>
-		<PackageProjectUrl>https://github.com/TheEightBot/Tycho</PackageProjectUrl>
-		<PackageTags>nosql database json</PackageTags>
-		<Title>Tycho DB - System.Text.Json Serializer</Title>
-		<Summary>Lightweight and highly opioninated NoSQL database written on top of SQLite</Summary>
-		<Description>Lightweight and highly opioninated NoSQL database written on top of SQLite</Description>
-	</PropertyGroup>
-	<ItemGroup>
-		<None Remove="System.Text.Json" />
-	</ItemGroup>
-	<ItemGroup>
-		<PackageReference Include="System.Text.Json" Version="7.0.3" />
-	</ItemGroup>
-	<ItemGroup>
-		<None Include="..\icon.png" PackagePath="icon.png" Pack="true" />
-	</ItemGroup>
-	<ItemGroup>
-	  <ProjectReference Include="..\Tycho.JsonSerializer\Tycho.JsonSerializer.csproj" />
-	</ItemGroup>
-</Project>
