--- Tycho.JsonSerializer/Tycho.JsonSerializer.csproj
@@ -1,21 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-			<TargetFrameworks>netstandard2.1;NET7.0</TargetFrameworks>
-			<LangVersion>Latest</LangVersion>
-			<PackageId>Tycho.JsonSerializer</PackageId>
-			<Authors>Eight-Bot, Inc.</Authors>
-			<Copyright>2023 Eight-Bot</Copyright>
-			<PackageIcon>icon.png</PackageIcon>
-			<Owners>Eight-Bot, Inc.</Owners>
-			<PackageProjectUrl>https://github.com/TheEightBot/Tycho</PackageProjectUrl>
-			<PackageTags>nosql database json</PackageTags>
-			<Title>Tycho DB - Newtonsoft JSON.NET Serializer</Title>
-			<Summary>Lightweight and highly opioninated NoSQL database written on top of SQLite</Summary>
-			<Description>Lightweight and highly opioninated NoSQL database written on top of SQLite</Description>
-    </PropertyGroup>
-
-	<ItemGroup>
-		<None Include="..\icon.png" PackagePath="icon.png" Pack="true" />
-	</ItemGroup>
-</Project>
