+++ TychoDB.Benchmarks/TychoDB.Benchmarks.csproj
@@ -1,0 +1,35 @@
+<Project Sdk="Microsoft.NET.Sdk">
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net8.0</TargetFramework>
+        <Configurations>Debug;Release;Encrypted</Configurations>
+    </PropertyGroup>
+    <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' " />
+    <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
+        <WarningLevel>4</WarningLevel>
+    </PropertyGroup>
+    <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Encrypted|AnyCPU' ">
+        <IntermediateOutputPath>obj\Release</IntermediateOutputPath>
+        <Optimize>true</Optimize>
+        <DefineConstants>ENCRYPTED;</DefineConstants>
+        <WarningLevel>4</WarningLevel>
+    </PropertyGroup>
+    <ItemGroup>
+        <None Remove="Benchmarks\" />
+        <None Remove="BenchmarkDotNet" />
+        <None Remove="sqlite-net-pcl" />
+    </ItemGroup>
+    <ItemGroup>
+        <Folder Include="Benchmarks\" />
+    </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="BenchmarkDotNet" Version="0.13.11" />
+        <PackageReference Include="sqlite-net-pcl" Version="1.8.116" />
+    </ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="../TychoDB.JsonSerializer.NewtonsoftJson/TychoDB.JsonSerializer.NewtonsoftJson.csproj" />
+        <ProjectReference Include="../TychoDB.JsonSerializer.SystemTextJson/TychoDB.JsonSerializer.SystemTextJson.csproj" />
+        <ProjectReference Include="../TychoDB.JsonSerializer/TychoDB.JsonSerializer.csproj" />
+        <ProjectReference Include="../TychoDB/TychoDB.csproj" />
+    </ItemGroup>
+</Project>
