+++ TychoDB.JsonSerializer.NewtonsoftJson/TychoDB.JsonSerializer.NewtonsoftJson.csproj
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project Sdk="MSBuild.Sdk.Extras/3.0.22">
+    <PropertyGroup>
+        <TargetFrameworks>netstandard2.1;NET8.0</TargetFrameworks>
+        <LangVersion>Latest</LangVersion>
+        <PackageId>TychoDB.JsonSerializer.NewtonsoftJson</PackageId>
+    </PropertyGroup>
+    <ItemGroup>
+        <PackageReference Include="Newtonsoft.Json"
+                          Version="13.0.3" />
+    </ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="../TychoDB.JsonSerializer/TychoDB.JsonSerializer.csproj" />
+    </ItemGroup>
+</Project>
