+++ TychoDB.JsonSerializer/IJsonSerializer.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.IO;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace TychoDB;
+
+public interface IJsonSerializer
+{
+    string DateTimeSerializationFormat { get; }
+
+    object Serialize<T>(T obj);
+
+    ValueTask<T> DeserializeAsync<T>(Stream stream, CancellationToken cancellationToken);
+}
