+++ TychoDB.UnitTests/TychoDB.UnitTests.csproj
@@ -1,0 +1,46 @@
+<Project Sdk="Microsoft.NET.Sdk">
+    <PropertyGroup>
+        <TargetFramework>net8.0</TargetFramework>
+        <IsPackable>false</IsPackable>
+        <Configurations>Debug;Release;Encrypted</Configurations>
+    </PropertyGroup>
+    <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
+        <DefineConstants>TRACE;DEBUG;NET;NET5_0;NETCOREAPP;</DefineConstants>
+        <WarningLevel>4</WarningLevel>
+    </PropertyGroup>
+    <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
+        <DebugType />
+        <WarningLevel>4</WarningLevel>
+    </PropertyGroup>
+    <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Encrypted|AnyCPU' ">
+        <IntermediateOutputPath>obj\Release\net5.0</IntermediateOutputPath>
+        <DebugType />
+        <Optimize>true</Optimize>
+        <OutputPath>bin\Encrypted\net5.0</OutputPath>
+        <DefineConstants>TRACE;RELEASE;NET;NET5_0;NETCOREAPP;ENCRYPTED;</DefineConstants>
+        <WarningLevel>4</WarningLevel>
+        <DocumentationFile />
+        <NoWarn>1701;1702</NoWarn>
+        <NoStdLib>true</NoStdLib>
+    </PropertyGroup>
+    <ItemGroup>
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0" />
+        <PackageReference Include="MSTest.TestAdapter" Version="3.1.1" />
+        <PackageReference Include="MSTest.TestFramework" Version="3.1.1" />
+        <PackageReference Include="coverlet.collector" Version="6.0.0">
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+            <PrivateAssets>all</PrivateAssets>
+        </PackageReference>
+        <PackageReference Include="FluentAssertions" Version="6.12.0" />
+        <PackageReference Include="PropertyChanged.Fody" Version="4.1.0" PrivateAssets="All" />
+    </ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="../TychoDB.JsonSerializer.NewtonsoftJson/TychoDB.JsonSerializer.NewtonsoftJson.csproj" />
+        <ProjectReference Include="../TychoDB.JsonSerializer.SystemTextJson/TychoDB.JsonSerializer.SystemTextJson.csproj" />
+        <ProjectReference Include="../TychoDB.JsonSerializer/TychoDB.JsonSerializer.csproj" />
+        <ProjectReference Include="../TychoDB/TychoDB.csproj" />
+    </ItemGroup>
+    <ItemGroup>
+        <None Remove="PropertyChanged.Fody" />
+    </ItemGroup>
+</Project>
